+++ datalad/interface/base.py
@@ -315,7 +315,7 @@ def update_docstring_with_parameters(func, params, prefix=None, suffix=None,
     description of its parameters. The Parameter spec needs to match
     the number and names of the callables arguments.
     """
-    from inspect import getargspec
+    from datalad.utils import getargspec
     # get the signature
     ndefaults = 0
     args, varargs, varkw, defaults = getargspec(func)
@@ -529,7 +529,7 @@ class Interface(object):
         # XXX needs safety check for name collisions
         # XXX allow for parser kwargs customization
         parser_kwargs = {}
-        from inspect import getargspec
+        from datalad.utils import getargspec
         # get the signature
         ndefaults = 0
         args, varargs, varkw, defaults = getargspec(cls.__call__)
@@ -579,7 +579,7 @@ class Interface(object):
     @classmethod
     def call_from_parser(cls, args):
         # XXX needs safety check for name collisions
-        from inspect import getargspec
+        from datalad.utils import getargspec
         argspec = getargspec(cls.__call__)
         if argspec[2] is None:
             # no **kwargs in the call receiver, pull argnames from signature
@@ -678,7 +678,7 @@ def get_allargs_as_kwargs(call, args, kwargs):
     resolvin the defaults for all kwargs that are not given in a kwargs
     dict
     """
-    from inspect import getargspec
+    from datalad.utils import getargspec
     argspec = getargspec(call)
     defaults = argspec.defaults
     nargs = len(argspec.args)