+++ datalad/plugin/tests/test_addurls.py
@@ -13,11 +13,13 @@ import json
 import logging
 import os
 import os.path as op
+import shutil
 import tempfile
 
 from mock import patch
 
 from six.moves import StringIO
+from six import text_type
 
 from datalad.api import addurls, Dataset, subdatasets
 import datalad.plugin.addurls as au
@@ -25,9 +27,11 @@ from datalad.support.exceptions import IncompleteResultsError
 from datalad.tests.utils import chpwd, slow, swallow_logs
 from datalad.tests.utils import assert_false, assert_true, assert_raises
 from datalad.tests.utils import assert_in, assert_re_in, assert_in_results
+from datalad.tests.utils import assert_not_in
 from datalad.tests.utils import assert_dict_equal
 from datalad.tests.utils import eq_, ok_exists
 from datalad.tests.utils import create_tree, with_tempfile, HTTPPath
+from datalad.tests.utils import with_tree
 from datalad.utils import get_tempfile_kwargs, rmtemp
 
 
@@ -316,36 +320,63 @@ def test_addurls_nonannex_repo(path):
     assert_in("not an annex repo", str(raised.exception))
 
 
+@with_tree({"in.csv": "linky,abcd\nhttps://datalad.org,f"})
+def test_addurls_unknown_placeholder(path):
+    ds = Dataset(path).create(force=True)
+    # Close but wrong URL placeholder
+    with assert_raises(IncompleteResultsError) as exc:
+        ds.addurls("in.csv", "{link}", "{abcd}", dry_run=True)
+    assert_in("linky", str(exc.exception))
+    # Close but wrong file name placeholder
+    with assert_raises(IncompleteResultsError) as exc:
+        ds.addurls("in.csv", "{linky}", "{abc}", dry_run=True)
+    assert_in("abcd", str(exc.exception))
+    # Out-of-bounds index.
+    with assert_raises(IncompleteResultsError) as exc:
+        ds.addurls("in.csv", "{linky}", "{3}", dry_run=True)
+    assert_in("index", str(exc.exception))
+
+    # Suggestions also work for automatic file name placeholders
+    with assert_raises(IncompleteResultsError) as exc:
+        ds.addurls("in.csv", "{linky}", "{_url_hostnam}", dry_run=True)
+    assert_in("_url_hostname", str(exc.exception))
+    # ... though if you whiff on the beginning prefix, we don't suggest
+    # anything because we decide to generate those fields based on detecting
+    # the prefix.
+    with assert_raises(IncompleteResultsError) as exc:
+        ds.addurls("in.csv", "{linky}", "{_uurl_hostnam}", dry_run=True)
+    assert_not_in("_url_hostname", str(exc.exception))
+
+
 @with_tempfile(mkdir=True)
 def test_addurls_dry_run(path):
     ds = Dataset(path).create(force=True)
 
-    with chpwd(path):
-        json_file = "links.json"
-        with open(json_file, "w") as jfh:
-            json.dump([{"url": "URL/a.dat", "name": "a", "subdir": "foo"},
-                       {"url": "URL/b.dat", "name": "b", "subdir": "bar"},
-                       {"url": "URL/c.dat", "name": "c", "subdir": "foo"}],
-                      jfh)
+    json_file = "links.json"
+    with open(op.join(ds.path, json_file), "w") as jfh:
+        json.dump([{"url": "URL/a.dat", "name": "a", "subdir": "foo"},
+                   {"url": "URL/b.dat", "name": "b", "subdir": "bar"},
+                   {"url": "URL/c.dat", "name": "c", "subdir": "foo"}],
+                  jfh)
 
-        ds.add(".", message="setup")
+    ds.add(".", message="setup")
 
-        with swallow_logs(new_level=logging.INFO) as cml:
-            ds.addurls(json_file,
-                       "{url}",
-                       "{subdir}//{_url_filename_root}",
-                       dry_run=True)
+    with swallow_logs(new_level=logging.INFO) as cml:
+        ds.addurls(json_file,
+                   "{url}",
+                   "{subdir}//{_url_filename_root}",
+                   dry_run=True)
 
-            for dir_ in ["foo", "bar"]:
-                assert_in("Would create a subdataset at {}".format(dir_),
-                          cml.out)
-            assert_in(
-                "Would download URL/a.dat to {}".format(
-                    os.path.join(path, "foo", "BASE")),
-                cml.out)
-
-            assert_in("Metadata: {}".format([u"name=a", u"subdir=foo"]),
+        for dir_ in ["foo", "bar"]:
+            assert_in("Would create a subdataset at {}".format(dir_),
                       cml.out)
+        assert_in(
+            "Would download URL/a.dat to {}".format(
+                os.path.join(path, "foo", "BASE")),
+            cml.out)
+
+        assert_in("Metadata: {}".format([u"name=a", u"subdir=foo"]),
+                  cml.out)
 
 
 @slow  # ~9s
@@ -387,51 +418,72 @@ class TestAddurls(object):
 
         n_annex_commits = get_annex_commit_counts()
 
-        with chpwd(path):
-            ds.addurls(self.json_file, "{url}", "{name}")
+        ds.addurls(self.json_file, "{url}", "{name}")
 
-            filenames = ["a", "b", "c"]
+        filenames = ["a", "b", "c"]
+        for fname in filenames:
+            ok_exists(op.join(ds.path, fname))
+
+        for (fname, meta), subdir in zip(ds.repo.get_metadata(filenames),
+                                         ["foo", "bar", "foo"]):
+            assert_dict_equal(meta,
+                              {"subdir": [subdir], "name": [fname]})
+
+        # Ignore this check if we're faking dates because that disables
+        # batch mode.
+        if not os.environ.get('DATALAD_FAKE__DATES'):
+            # We should have two new commits on the git-annex: one for the
+            # added urls and one for the added metadata.
+            eq_(n_annex_commits + 2, get_annex_commit_counts())
+
+        # Add to already existing links, overwriting.
+        with swallow_logs(new_level=logging.DEBUG) as cml:
+            ds.addurls(self.json_file, "{url}", "{name}",
+                       ifexists="overwrite")
             for fname in filenames:
-                ok_exists(fname)
+                assert_in("Removing {}".format(os.path.join(path, fname)),
+                          cml.out)
 
-            for (fname, meta), subdir in zip(ds.repo.get_metadata(filenames),
-                                             ["foo", "bar", "foo"]):
-                assert_dict_equal(meta,
-                                  {"subdir": [subdir], "name": [fname]})
+        # Add to already existing links, skipping.
+        assert_in_results(
+            ds.addurls(self.json_file, "{url}", "{name}", ifexists="skip"),
+            action="addurls",
+            status="notneeded")
 
-            # Ignore this check if we're faking dates because that disables
-            # batch mode.
-            if not os.environ.get('DATALAD_FAKE__DATES'):
-                # We should have two new commits on the git-annex: one for the
-                # added urls and one for the added metadata.
-                eq_(n_annex_commits + 2, get_annex_commit_counts())
+        # Add to already existing links works, as long content is the same.
+        ds.addurls(self.json_file, "{url}", "{name}")
 
-            # Add to already existing links, overwriting.
-            with swallow_logs(new_level=logging.DEBUG) as cml:
-                ds.addurls(self.json_file, "{url}", "{name}",
-                           ifexists="overwrite")
-                for fname in filenames:
-                    assert_in("Removing {}".format(os.path.join(path, fname)),
-                              cml.out)
+        # But it fails if something has changed.
+        ds.unlock("a")
+        with open(op.join(ds.path, "a"), "w") as ofh:
+            ofh.write("changed")
+        ds.add("a")
 
-            # Add to already existing links, skipping.
-            assert_in_results(
-                ds.addurls(self.json_file, "{url}", "{name}", ifexists="skip"),
-                action="addurls",
-                status="notneeded")
+        assert_raises(IncompleteResultsError,
+                      ds.addurls,
+                      self.json_file, "{url}", "{name}")
 
-            # Add to already existing links works, as long content is the same.
-            ds.addurls(self.json_file, "{url}", "{name}")
+    @with_tempfile(mkdir=True)
+    def test_addurls_unbound_dataset(self, path):
+        ds = Dataset(path).create(force=True)
 
-            # But it fails if something has changed.
-            ds.unlock("a")
-            with open("a", "w") as ofh:
-                ofh.write("changed")
-            ds.add("a")
+        def check(subpath, dataset_arg, url_file):
+            subdir = op.join(path, subpath)
+            os.mkdir(subdir)
+            with chpwd(subdir):
+                shutil.copy(self.json_file, "in.json")
+                addurls(dataset_arg, url_file, "{url}", "{name}")
+                # Files specified in the CSV file are always relative to the
+                # dataset.
+                for fname in ["a", "b", "c"]:
+                    ok_exists(op.join(ds.path, fname))
 
-            assert_raises(IncompleteResultsError,
-                          ds.addurls,
-                          self.json_file, "{url}", "{name}")
+        # The input file is relative to the current working directory, as
+        # with other commands.
+        check("subdir0", None, "in.json")
+        # The input file (without a leading "./") is relative to the dataset if
+        # any dataset argument is given, even a string.
+        check("subdir1", ds.path, op.join("subdir1", "in.json"))
 
     @with_tempfile(mkdir=True)
     def test_addurls_create_newdataset(self, path):
@@ -444,73 +496,68 @@ class TestAddurls(object):
     def test_addurls_subdataset(self, path):
         ds = Dataset(path).create(force=True)
 
-        with chpwd(path):
-            for save in True, False:
-                label = "save" if save else "nosave"
-                hexsha_before = ds.repo.get_hexsha()
-                ds.addurls(self.json_file, "{url}",
-                           "{subdir}-" + label + "//{name}",
-                           save=save)
-                hexsha_after = ds.repo.get_hexsha()
+        for save in True, False:
+            label = "save" if save else "nosave"
+            hexsha_before = ds.repo.get_hexsha()
+            ds.addurls(self.json_file, "{url}",
+                       "{subdir}-" + label + "//{name}",
+                       save=save)
+            hexsha_after = ds.repo.get_hexsha()
 
-                for fname in ["foo-{}/a", "bar-{}/b", "foo-{}/c"]:
-                    ok_exists(fname.format(label))
+            for fname in ["foo-{}/a", "bar-{}/b", "foo-{}/c"]:
+                ok_exists(op.join(ds.path, fname.format(label)))
 
-                assert_true(save ^ (hexsha_before == hexsha_after))
-                assert_true(save ^ ds.repo.dirty)
+            assert_true(save ^ (hexsha_before == hexsha_after))
+            assert_true(save ^ ds.repo.dirty)
 
-            # Now save the "--nosave" changes and check that we have
-            # all the subdatasets.
-            ds.add(".")
-            eq_(set(subdatasets(ds, recursive=True,
-                                result_xfm="relpaths")),
-                {"foo-save", "bar-save", "foo-nosave", "bar-nosave"})
+        # Now save the "--nosave" changes and check that we have
+        # all the subdatasets.
+        ds.add(".")
+        eq_(set(subdatasets(ds, recursive=True,
+                            result_xfm="relpaths")),
+            {"foo-save", "bar-save", "foo-nosave", "bar-nosave"})
 
-            # We don't try to recreate existing subdatasets.
-            with swallow_logs(new_level=logging.DEBUG) as cml:
-                ds.addurls(self.json_file, "{url}", "{subdir}-nosave//{name}")
-                assert_in("Not creating subdataset at existing path", cml.out)
+        # We don't try to recreate existing subdatasets.
+        with swallow_logs(new_level=logging.DEBUG) as cml:
+            ds.addurls(self.json_file, "{url}", "{subdir}-nosave//{name}")
+            assert_in("Not creating subdataset at existing path", cml.out)
 
     @with_tempfile(mkdir=True)
     def test_addurls_repindex(self, path):
         ds = Dataset(path).create(force=True)
 
-        with chpwd(path):
-            with assert_raises(IncompleteResultsError) as raised:
-                ds.addurls(self.json_file, "{url}", "{subdir}")
-            assert_in("There are file name collisions", str(raised.exception))
+        with assert_raises(IncompleteResultsError) as raised:
+            ds.addurls(self.json_file, "{url}", "{subdir}")
+        assert_in("There are file name collisions", str(raised.exception))
 
-            ds.addurls(self.json_file, "{url}", "{subdir}-{_repindex}")
+        ds.addurls(self.json_file, "{url}", "{subdir}-{_repindex}")
 
-            for fname in ["foo-0", "bar-0", "foo-1"]:
-                ok_exists(fname)
+        for fname in ["foo-0", "bar-0", "foo-1"]:
+            ok_exists(op.join(ds.path, fname))
 
     @with_tempfile(mkdir=True)
     def test_addurls_url_parts(self, path):
         ds = Dataset(path).create(force=True)
-        with chpwd(path):
-            ds.addurls(self.json_file, "{url}", "{_url0}/{_url_basename}")
+        ds.addurls(self.json_file, "{url}", "{_url0}/{_url_basename}")
 
-            for fname in ["udir/a.dat", "udir/b.dat", "udir/c.dat"]:
-                ok_exists(fname)
+        for fname in ["udir/a.dat", "udir/b.dat", "udir/c.dat"]:
+            ok_exists(op.join(ds.path, fname))
 
     @with_tempfile(mkdir=True)
     def test_addurls_url_filename(self, path):
         ds = Dataset(path).create(force=True)
-        with chpwd(path):
-            ds.addurls(self.json_file, "{url}", "{_url0}/{_url_filename}")
-            for fname in ["udir/a.dat", "udir/b.dat", "udir/c.dat"]:
-                ok_exists(fname)
+        ds.addurls(self.json_file, "{url}", "{_url0}/{_url_filename}")
+        for fname in ["udir/a.dat", "udir/b.dat", "udir/c.dat"]:
+            ok_exists(op.join(ds.path, fname))
 
     @with_tempfile(mkdir=True)
     def test_addurls_url_filename_fail(self, path):
         ds = Dataset(path).create(force=True)
-        with chpwd(path):
-            assert_raises(IncompleteResultsError,
-                          ds.addurls,
-                          self.json_file,
-                          "{url}/nofilename/",
-                          "{_url0}/{_url_filename}")
+        assert_raises(IncompleteResultsError,
+                      ds.addurls,
+                      self.json_file,
+                      "{url}/nofilename/",
+                      "{_url0}/{_url_filename}")
 
     @with_tempfile(mkdir=True)
     def test_addurls_metafail(self, path):
@@ -523,14 +570,14 @@ class TestAddurls(object):
             for i in fn("wreaking-havoc-and-such", **kwargs):
                 yield i
 
-        with chpwd(path), patch.object(ds.repo, 'set_metadata', set_meta):
+        with patch.object(ds.repo, 'set_metadata', set_meta):
             with assert_raises(IncompleteResultsError):
                 ds.addurls(self.json_file, "{url}", "{name}")
 
     @with_tempfile(mkdir=True)
     def test_addurls_dropped_urls(self, path):
         ds = Dataset(path).create(force=True)
-        with chpwd(path), swallow_logs(new_level=logging.WARNING) as cml:
+        with swallow_logs(new_level=logging.WARNING) as cml:
             ds.addurls(self.json_file, "", "{subdir}//{name}")
             assert_re_in(r".*Dropped [0-9]+ row\(s\) that had an empty URL",
                          str(cml.out))
@@ -574,3 +621,24 @@ class TestAddurls(object):
              op.join("bar", "adir", "bar-again", "other-ds")})
         ok_exists(os.path.join(
             ds.path, "foo", "adir", "foo-again", "other-ds", "bdir", "a"))
+
+    @with_tree({"in": ""})
+    def test_addurls_invalid_input(self, path):
+        ds = Dataset(path).create(force=True)
+        in_file = op.join(path, "in")
+        for in_type in ["csv", "json"]:
+            with assert_raises(IncompleteResultsError) as exc:
+                ds.addurls(in_file, "{url}", "{name}", input_type=in_type)
+            assert_in("Failed to read", text_type(exc.exception))
+
+    @with_tree({"in.csv": "url,name,subdir",
+                "in.json": "[]"})
+    def test_addurls_no_rows(self, path):
+        ds = Dataset(path).create(force=True)
+        for fname in ["in.csv", "in.json"]:
+            with swallow_logs(new_level=logging.WARNING) as cml:
+                assert_in_results(
+                    ds.addurls(fname, "{url}", "{name}"),
+                    action="addurls",
+                    status="notneeded")
+                cml.assert_logged("No rows", regex=False)