+++ datalad/plugin/wtf.py
@@ -16,6 +16,8 @@ import os.path as op
 from functools import partial
 from collections import OrderedDict
 
+from six import PY2
+
 from datalad.interface.base import Interface
 from datalad.interface.base import build_doc
 from datalad.utils import (
@@ -112,26 +114,12 @@ def _describe_annex():
 def _describe_system():
     import platform as pl
     from datalad import get_encoding_info
-
-    if hasattr(pl, 'dist'):
-        dist = pl.dist()
-    else:
-        # Python 3.8 removed .dist but recommended "distro" is slow, so we
-        # try it only if needed
-        try:
-            import distro
-            dist = distro.linux_distribution(full_distribution_name=False)
-        except ImportError:
-            lgr.info(
-                "Please install 'distro' package to obtain distribution information"
-            )
-            dist = tuple()
-        except Exception as exc:
-            lgr.warning(
-                "No distribution information will be provided since 'distro' "
-                "fails to import/run: %s", exc_str(exc)
-            )
-            dist = tuple()
+    from datalad.utils import get_linux_distribution
+    try:
+        dist = get_linux_distribution()
+    except Exception as exc:
+        lgr.warning("Failed to get distribution information: %s", exc_str(exc))
+        dist = tuple()
 
     return {
         'type': os.name,
@@ -365,7 +353,7 @@ class WTF(Interface):
         infos = OrderedDict()
         res = get_status_dict(
             action='wtf',
-            path=ds.path if ds else op.abspath(op.curdir),
+            path=ds.path if ds else assure_unicode(op.abspath(op.curdir)),
             type='dataset' if ds else 'directory',
             status='ok',
             logger=lgr,
@@ -397,7 +385,7 @@ class WTF(Interface):
                 'pyperclip', msg="It is needed to be able to use clipboard")
             import pyperclip
             report = _render_report(res)
-            pyperclip.copy(assure_bytes(report))
+            pyperclip.copy(report)
             ui.message("WTF information of length %s copied to clipboard"
                        % len(report))
         yield res
@@ -406,7 +394,8 @@ class WTF(Interface):
     @staticmethod
     def custom_result_renderer(res, **kwargs):
         from datalad.ui import ui
-        ui.message(_render_report(res))
+        out = _render_report(res)
+        ui.message(assure_bytes(out) if PY2 else out)
 
 
 def _render_report(res):