+++ datalad/support/annexrepo.py
@@ -120,6 +120,7 @@ class AnnexRepo(GitRepo, RepoInterface):
     # 6.20180913 -- annex fixes all known to us issues for v6
     GIT_ANNEX_MIN_VERSION = '6.20180913'
     git_annex_version = None
+    supports_direct_mode = None
 
     # Class wide setting to allow insecure URLs. Used during testing, since
     # git annex 6.20180626 those will by default be not allowed for security
@@ -257,6 +258,8 @@ class AnnexRepo(GitRepo, RepoInterface):
             else:
                 raise InvalidAnnexRepositoryError("No annex found at %s." % self.path)
 
+        # Ensure self.supports_direct_mode is set.
+        self.check_direct_mode_support()
         self._direct_mode = None  # we don't know yet
 
         # If we are in direct mode already, we need to make
@@ -278,7 +281,8 @@ class AnnexRepo(GitRepo, RepoInterface):
             # Note: If 'annex.version' is missing in .git/config for some
             # reason, we need to try to set direct mode:
             repo_version = self.config.getint("annex", "version")
-            if (repo_version is None) or (repo_version < 6):
+            if self.supports_direct_mode and \
+               (repo_version is None or repo_version < 6):
                 lgr.debug("Switching to direct mode (%s)." % self)
                 self.set_direct_mode()
             else:
@@ -778,6 +782,22 @@ class AnnexRepo(GitRepo, RepoInterface):
         elif ver < cls.GIT_ANNEX_MIN_VERSION:
             raise OutdatedExternalDependency(ver_present=ver, **exc_kwargs)
         cls.git_annex_version = ver
+
+    @classmethod
+    def check_direct_mode_support(cls):
+        """Does git-annex version support direct mode?
+
+        The result is cached at `cls.supports_direct_mode`.
+
+        Returns
+        -------
+        bool
+        """
+        if cls.supports_direct_mode is None:
+            if cls.git_annex_version is None:
+                cls._check_git_annex_version()
+            cls.supports_direct_mode = cls.git_annex_version <= "7.20190819"
+        return cls.supports_direct_mode
 
     @staticmethod
     def get_size_from_key(key):
@@ -1383,7 +1403,7 @@ class AnnexRepo(GitRepo, RepoInterface):
 
         if not key:
             expected_downloads, fetch_files = self._get_expected_files(
-                files, ['--not', '--in', 'here'],
+                files, ['--not', '--in', '.'],
                 merge_annex_branches=False  # interested only in local info
             )
         else:
@@ -2751,7 +2771,7 @@ class AnnexRepo(GitRepo, RepoInterface):
                 args.append('-)')
 
             if with_content_only:
-                args.extend(['--in', 'here'])
+                args.extend(['--in', '.'])
         out, err = self._run_annex_command(
             'find', annex_options=args, merge_annex_branches=False
         )
@@ -3296,7 +3316,7 @@ class AnnexRepo(GitRepo, RepoInterface):
         # is the verb (copy, copy) or (get, put) and remote ('here', remote)?
         if '--key' not in options:
             expected_copys, copy_files = self._get_expected_files(
-                files, ['--in', 'here', '--not', '--in', remote])
+                files, ['--in', '.', '--not', '--in', remote])
         else:
             copy_files = files
             assert(len(files) == 1)