+++ datalad/tests/utils.py
@@ -1097,6 +1097,30 @@ def skip_direct_mode(func, method='raise'):
 
 
 @optional_args
+def skip_if_no_direct_mode(func, other_cond=True, method='raise'):
+    """Skip test if git-annex version does not support direct mode.
+
+    Parameters
+    ----------
+    func : function
+    other_cond : bool, optional
+        Skip if annex does not support direct mode AND this value is true.
+    method : str, optional
+        Passed to `skip_if`.
+    """
+    from datalad.support.annexrepo import AnnexRepo
+    unsupported = not AnnexRepo.check_direct_mode_support()
+
+    @skip_if(unsupported and other_cond,
+             msg="Direct mode unsupported by git-annex version",
+             method=method)
+    @wraps(func)
+    def newfunc(*args, **kwargs):
+        return func(*args, **kwargs)
+    return newfunc
+
+
+@optional_args
 def assert_cwd_unchanged(func, ok_to_chdir=False):
     """Decorator to test whether the current working directory remains unchanged
 
@@ -1548,6 +1572,24 @@ def set_date(timestamp):
                      "DATALAD_FAKE__DATES": "0"}):
         yield
 
+
+@contextmanager
+def set_annex_version(version):
+    """Override the git-annex version.
+
+    This temporarily masks the git-annex version present in external_versions
+    and make AnnexRepo forget its cached version information.
+    """
+    from datalad.support.annexrepo import AnnexRepo
+    ar_vers = AnnexRepo.git_annex_version
+    with patch.dict(
+            "datalad.support.annexrepo.external_versions._versions",
+            {"cmd:annex": version}):
+        try:
+            AnnexRepo.git_annex_version = None
+            yield
+        finally:
+            AnnexRepo.git_annex_version = ar_vers
 
 #
 # Test tags